+++ interop/kotlinx-metadata/src/main/kotlin/com/squareup/kotlinpoet/metadata/KotlinPoetMetadata.kt
@@ -35,28 +35,45 @@ import kotlinx.metadata.jvm.Metadata
 @Target(CLASS, FUNCTION, PROPERTY)
 public annotation class KotlinPoetMetadataPreview
 
-/** @return a new [KmClass] representation of the Kotlin metadata for [this] class. */
+/**
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+ * @return a new [KmClass] representation of the Kotlin metadata for [this] class.
+ */
 @KotlinPoetMetadataPreview
-public fun KClass<*>.toKmClass(): KmClass = java.toKmClass()
+public fun KClass<*>.toKmClass(lenient: Boolean): KmClass = java.toKmClass(lenient)
 
-/** @return a new [KmClass] representation of the Kotlin metadata for [this] class. */
+/**
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+ * @return a new [KmClass] representation of the Kotlin metadata for [this] class.
+ */
 @KotlinPoetMetadataPreview
-public fun Class<*>.toKmClass(): KmClass = readMetadata(::getAnnotation).toKmClass()
+public fun Class<*>.toKmClass(lenient: Boolean): KmClass = readMetadata(::getAnnotation).toKmClass(lenient)
 
-/** @return a new [KmClass] representation of the Kotlin metadata for [this] type. */
+/**
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+ * @return a new [KmClass] representation of the Kotlin metadata for [this] type.
+ */
 @KotlinPoetMetadataPreview
-public fun TypeElement.toKmClass(): KmClass = readMetadata(::getAnnotation).toKmClass()
+public fun TypeElement.toKmClass(lenient: Boolean): KmClass = readMetadata(::getAnnotation).toKmClass(lenient)
 
+/**
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+ */
 @KotlinPoetMetadataPreview
-public fun Metadata.toKmClass(): KmClass {
-  return toKotlinClassMetadata<KotlinClassMetadata.Class>()
+public fun Metadata.toKmClass(lenient: Boolean): KmClass {
+  return toKotlinClassMetadata<KotlinClassMetadata.Class>(lenient)
     .kmClass
 }
 
+/**
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+ */
 @KotlinPoetMetadataPreview
-public inline fun <reified T : KotlinClassMetadata> Metadata.toKotlinClassMetadata(): T {
+public inline fun <reified T : KotlinClassMetadata> Metadata.toKotlinClassMetadata(
+  lenient: Boolean,
+): T {
   val expectedType = T::class
-  val metadata = readKotlinClassMetadata()
+  val metadata = readKotlinClassMetadata(lenient)
   return when (expectedType) {
     KotlinClassMetadata.Class::class -> {
       check(metadata is KotlinClassMetadata.Class)
@@ -82,30 +99,20 @@ public inline fun <reified T : KotlinClassMetadata> Metadata.toKotlinClassMetada
  * Returns the [KotlinClassMetadata] this represents. In general you should only use this function
  * when you don't know what the underlying [KotlinClassMetadata] subtype is, otherwise you should
  * use one of the more direct functions like [toKmClass].
+ *
+ * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
  */
 @KotlinPoetMetadataPreview
-public fun Metadata.readKotlinClassMetadata(): KotlinClassMetadata {
-  val metadata = KotlinClassMetadata.readStrict(asClassHeader())
-  checkNotNull(metadata) {
-    "Could not parse metadata! Try bumping kotlinpoet and/or kotlinx-metadata version."
+public fun Metadata.readKotlinClassMetadata(lenient: Boolean): KotlinClassMetadata {
+  return if (lenient) {
+    KotlinClassMetadata.readLenient(this)
+  } else {
+    KotlinClassMetadata.readStrict(this)
   }
-  return metadata
 }
 
 private inline fun readMetadata(lookup: ((Class<Metadata>) -> Metadata?)): Metadata {
   return checkNotNull(lookup.invoke(Metadata::class.java)) {
     "No Metadata annotation found! Must be Kotlin code built with the standard library on the classpath."
   }
-}
-
-private fun Metadata.asClassHeader(): Metadata {
-  return Metadata(
-    kind = kind,
-    metadataVersion = metadataVersion,
-    data1 = data1,
-    data2 = data2,
-    extraString = extraString,
-    packageName = packageName,
-    extraInt = extraInt,
-  )
 }